feat: extract PreviousEnvironmentHealthCommitStatus into its own controller
Move the previous environment health checking and commit status creation
logic from the PromotionStrategy controller into a new dedicated
PreviousEnvironmentHealthCommitStatus controller. This decouples the
linear environment gating logic from the PromotionStrategy reconciler,
allowing for different gating strategies to be implemented independently.

The new controller:
- Watches PromotionStrategy and ChangeTransferPolicy resources
- Evaluates previous environment health (hydrator status, promotion
  completion, commit time, and active commit statuses)
- Creates/updates CommitStatus resources on the current environment's
  proposed commit reflecting the previous environment's health

The PromotionStrategy controller retains the logic for adding the
PreviousEnvironmentCommitStatusKey selector to CTP proposed commit
statuses, as that is part of CTP spec configuration.
